added dpdk as a submodule

+++ build.sh
@@ -3,13 +3,17 @@
 # TODO: this should probably be a makefile
 # TODO: install target
 (
-# Compile dpdk
-cd ./dpdk
+cd deps/dpdk
+if [[ ! -e Makefile ]]
+then
+	echo "ERROR: dpdk submodule not initialized"
+	echo "Please run git submodule update --init"
+	exit 1
+fi
 make -j 8 install T=x86_64-native-linuxapp-gcc
-
-# Compile pktgen
+cd ../../
+cd setup-scripts
+source ./helpers/setup-vars-pktgen-dpdk.sh
 cd ../
-export RTE_SDK=$PWD/dpdk
-export RTE_TARGET=x86_64-native-linuxapp-gcc
 make
 )
